moved the distributed transaction id in the server module
