+++ distributed/src/main/java/com/orientechnologies/orient/server/distributed/impl/ODatabaseDocumentDistributed.java
@@ -47,7 +47,7 @@ import com.orientechnologies.orient.server.distributed.impl.metadata.OClassDistr
 import com.orientechnologies.orient.server.distributed.impl.metadata.OSharedContextDistributed;
 import com.orientechnologies.orient.server.distributed.impl.task.ONewSQLCommandTask;
 import com.orientechnologies.orient.server.distributed.impl.task.ORunQueryExecutionPlanTask;
-import com.orientechnologies.orient.server.distributed.impl.task.transaction.OTransactionId;
+import com.orientechnologies.orient.server.distributed.OTransactionId;
 import com.orientechnologies.orient.server.distributed.task.*;
 import com.orientechnologies.orient.server.hazelcast.OHazelcastPlugin;
 import com.orientechnologies.orient.server.plugin.OServerPluginInfo;
@@ -454,12 +454,7 @@ public class ODatabaseDocumentDistributed extends ODatabaseDocumentEmbedded {
     try {
       internalBegin2pc(txContext, local);
       txContext.setStatus(SUCCESS);
-      if (local) {
-        // No need to check the ID
-        localDistributedDatabase.registerTxContext(requestId, null, txContext);
-      } else {
-        localDistributedDatabase.registerTxContext(requestId, id, txContext);
-      }
+      register(requestId, id, local, localDistributedDatabase, txContext);
     } catch (OConcurrentCreateException ex) {
       if (retryCount >= 0 && retryCount < getConfiguration().getValueAsInteger(DISTRIBUTED_CONCURRENT_TX_MAX_AUTORETRY)) {
         if (ex.getExpectedRid().getClusterPosition() > ex.getActualRid().getClusterPosition()) {
@@ -471,7 +466,7 @@ public class ODatabaseDocumentDistributed extends ODatabaseDocumentEmbedded {
         }
       }
       txContext.setStatus(FAILED);
-      localDistributedDatabase.registerTxContext(requestId, id, txContext);
+      register(requestId, id, local, localDistributedDatabase, txContext);
       throw ex;
     } catch (OConcurrentModificationException ex) {
       if (retryCount >= 0 && retryCount < getConfiguration().getValueAsInteger(DISTRIBUTED_CONCURRENT_TX_MAX_AUTORETRY)) {
@@ -484,7 +479,7 @@ public class ODatabaseDocumentDistributed extends ODatabaseDocumentEmbedded {
         }
       }
       txContext.setStatus(FAILED);
-      localDistributedDatabase.registerTxContext(requestId, id, txContext);
+      register(requestId, id, local, localDistributedDatabase, txContext);
       throw ex;
     } catch (ORecordNotFoundException e) {
       // This error can happen only in deserialization before locks happen, no need to unlock
@@ -492,22 +487,32 @@ public class ODatabaseDocumentDistributed extends ODatabaseDocumentEmbedded {
         return false;
       }
       txContext.setStatus(FAILED);
-      localDistributedDatabase.registerTxContext(requestId, id, txContext);
+      register(requestId, id, local, localDistributedDatabase, txContext);
       throw e;
     } catch (ODistributedRecordLockedException | ODistributedKeyLockedException ex) {
       /// ?? do i've to save this state as well ?
       txContext.setStatus(TIMEDOUT);
-      getStorageDistributed().getLocalDistributedDatabase().registerTxContext(requestId, id, txContext);
+      register(requestId, id, local, localDistributedDatabase, txContext);
       throw ex;
     } catch (ORecordDuplicatedException ex) {
       txContext.setStatus(FAILED);
-      localDistributedDatabase.registerTxContext(requestId, id, txContext);
+      register(requestId, id, local, localDistributedDatabase, txContext);
       throw ex;
     } catch (OLowDiskSpaceException ex) {
       distributedManager.setDatabaseStatus(getLocalNodeName(), getName(), ODistributedServerManager.DB_STATUS.OFFLINE);
       throw ex;
     }
     return true;
+  }
+
+  private void register(ODistributedRequestId requestId, OTransactionId id, boolean local,
+      ODistributedDatabase localDistributedDatabase, ONewDistributedTxContextImpl txContext) {
+    if (local) {
+      // No need to check the ID
+      localDistributedDatabase.registerTxContext(requestId, Optional.empty(), txContext);
+    } else {
+      localDistributedDatabase.registerTxContext(requestId, Optional.of(id), txContext);
+    }
   }
 
   /**
